+++ gdsfactory/tests/test_transition.py
@@ -4,7 +4,8 @@ from phidl.path import transition
 import gdsfactory as gf
 
 
-def test_transition():
+def test_transition_unamed_fails():
+    """raises error when transitioning un-named cross_sections"""
     with pytest.raises(ValueError):
 
         path = gf.Path()
@@ -26,5 +27,23 @@ def test_transition():
         transition(X, x2)
 
 
+def test_transition_ports():
+    width1 = 0.5
+    width2 = 1.0
+    x1 = gf.cross_section.strip(width=width1)
+    x2 = gf.cross_section.strip(width=width2)
+    xt = gf.path.transition(cross_section1=x1, cross_section2=x2, width_type="linear")
+    path = gf.path.straight(length=5)
+    c = gf.path.extrude(path, xt)
+    assert c.ports["o1"].cross_section.info["width"] == width1
+    assert c.ports["o2"].cross_section.info["width"] == width2
+
+
 if __name__ == "__main__":
-    test_transition()
+    # test_transition_ports()
+    # x1 = gf.cross_section.strip(width=0.5)
+    # x2 = gf.cross_section.strip(width=1.0)
+    # xt = gf.path.transition(cross_section1=x1, cross_section2=x2, width_type="linear")
+    # path = gf.path.straight(length=5)
+    # c = gf.path.extrude(path, xt)
+    test_transition_ports()